kiosk/requirements.txt
tkinter
requests
werkzeug
nfcpy
